build dockerfile taotao-cloud-gateway taotao-cloud-monitor taotao-cloud-file  taotao-cloud-job  taotao-cloud-auth taotao-cloud-sys  taotao-cloud-xxljob

+++ taotao-cloud-microservice/taotao-cloud-business/taotao-cloud-payment/taotao-cloud-payment-biz/src/main/java/com/taotao/cloud/payment/biz/jeepay/service/impl/IsvInfoService.java
@@ -1,34 +1,35 @@
 /*
- * Copyright (c) 2020-2030, Shuigedeng (981376577@qq.com & https://blog.taotaocloud.top/).
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
+ * Copyright (c) 2021-2031, 河北计全科技有限公司 (https://www.jeequan.com & jeequan@126.com).
+ * <p>
+ * Licensed under the GNU LESSER GENERAL PUBLIC LICENSE 3.0;
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- *
- *      https://www.apache.org/licenses/LICENSE-2.0
- *
+ * <p>
+ * http://www.gnu.org/licenses/lgpl.html
+ * <p>
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
 package com.taotao.cloud.payment.biz.jeepay.service.impl;
 
 import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
-import com.taotao.cloud.payment.biz.jeepay.core.constants.CS;
-import com.taotao.cloud.payment.biz.jeepay.core.entity.IsvInfo;
-import com.taotao.cloud.payment.biz.jeepay.core.entity.MchInfo;
-import com.taotao.cloud.payment.biz.jeepay.core.entity.PayInterfaceConfig;
-import com.taotao.cloud.payment.biz.jeepay.core.exception.BizException;
-import com.taotao.cloud.payment.biz.jeepay.service.mapper.IsvInfoMapper;
+import com.jeequan.jeepay.core.constants.CS;
+import com.jeequan.jeepay.core.entity.IsvInfo;
+import com.jeequan.jeepay.core.entity.MchInfo;
+import com.jeequan.jeepay.core.entity.PayInterfaceConfig;
+import com.jeequan.jeepay.core.exception.BizException;
+import com.jeequan.jeepay.service.mapper.IsvInfoMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 /**
+ * <p>
  * 服务商信息表 服务实现类
+ * </p>
  *
  * @author [mybatis plus generator]
  * @since 2021-04-27
@@ -36,14 +37,11 @@ import org.springframework.transaction.annotation.Transactional;
 @Service
 public class IsvInfoService extends ServiceImpl<IsvInfoMapper, IsvInfo> {
 
-    @Autowired
-    private MchInfoService mchInfoService;
+    @Autowired private MchInfoService mchInfoService;
 
-    @Autowired
-    private IsvInfoService isvInfoService;
+    @Autowired private IsvInfoService isvInfoService;
 
-    @Autowired
-    private PayInterfaceConfigService payInterfaceConfigService;
+    @Autowired private PayInterfaceConfigService payInterfaceConfigService;
 
     @Transactional
     public void removeByIsvNo(String isvNo) {
@@ -54,8 +52,7 @@ public class IsvInfoService extends ServiceImpl<IsvInfoMapper, IsvInfo> {
         }
 
         // 1.查询当前服务商下是否存在商户
-        long mchCount = mchInfoService.count(
-                MchInfo.gw().eq(MchInfo::getIsvNo, isvNo).eq(MchInfo::getType, CS.MCH_TYPE_ISVSUB));
+        int mchCount = mchInfoService.count(MchInfo.gw().eq(MchInfo::getIsvNo, isvNo).eq(MchInfo::getType, CS.MCH_TYPE_ISVSUB));
         if (mchCount > 0) {
             throw new BizException("该服务商下存在商户，不可删除");
         }
@@ -63,7 +60,8 @@ public class IsvInfoService extends ServiceImpl<IsvInfoMapper, IsvInfo> {
         // 2.删除当前服务商支付接口配置参数
         payInterfaceConfigService.remove(PayInterfaceConfig.gw()
                 .eq(PayInterfaceConfig::getInfoId, isvNo)
-                .eq(PayInterfaceConfig::getInfoType, CS.INFO_TYPE_ISV));
+                .eq(PayInterfaceConfig::getInfoType, CS.INFO_TYPE_ISV)
+        );
 
         // 3.删除该服务商
         boolean remove = isvInfoService.removeById(isvNo);
